Refactor components and move canvas logic
Simplified DashboardComponent by removing canvas rendering and configuration fetching logic. Moved this functionality to EnvironmentGatewaysComponent, which now includes a Bootstrap accordion with buttons and a canvas. Created a new GatewayComponent with a basic template and unit tests.
