+++ src/DevServerClient/src/app/Pages/admin/dashboard/dashboard.component.html
@@ -1,8 +1,1 @@
 <p>dashboard works!</p>
-
-<button class="btn btn-primary" (click)="getConfig()">Get Config</button>
-<button class="btn btn-primary ms-2" (click)="getSlotOne()">Slot One</button>
-
-<div class="canvas-container">
-  <canvas #canvas width="800" height="2000"></canvas>
-</div>
