+++ src/DevServerClient/src/app/Pages/admin/environment-gateways/components/gateway/gateway.component.html
@@ -1,0 +1,1 @@
+<p>gateway works!</p>
