+++ src/DevServerClient/src/app/Pages/admin/environment-gateways/components/gateway/gateway.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-gateway',
+  imports: [],
+  templateUrl: './gateway.component.html',
+  styleUrl: './gateway.component.css'
+})
+export class GatewayComponent {
+
+}
