+++ src/DevServerClient/src/app/Pages/admin/environment-gateways/environment-gateways-page.component.html
@@ -1,1 +1,94 @@
-<p>environment-gateways works!</p>
+<div class="accordion" id="accordionExample">
+  <div class="accordion-item">
+    <h2 class="accordion-header">
+      <button
+        class="accordion-button"
+        type="button"
+        data-bs-toggle="collapse"
+        data-bs-target="#collapseOne"
+        aria-expanded="true"
+        aria-controls="collapseOne"
+      >
+        Accordion Item #1
+      </button>
+    </h2>
+    <div
+      id="collapseOne"
+      class="accordion-collapse collapse show"
+      data-bs-parent="#accordionExample"
+    >
+      <div class="accordion-body">
+        <button class="btn btn-primary" (click)="getConfig()">
+          Get Config
+        </button>
+        <button class="btn btn-primary ms-2" (click)="getSlotOne()">
+          Slot One
+        </button>
+
+        <div class="canvas-container">
+          <canvas #canvas width="800" height="1000"></canvas>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="accordion-item">
+    <h2 class="accordion-header">
+      <button
+        class="accordion-button collapsed"
+        type="button"
+        data-bs-toggle="collapse"
+        data-bs-target="#collapseTwo"
+        aria-expanded="false"
+        aria-controls="collapseTwo"
+      >
+        Accordion Item #2
+      </button>
+    </h2>
+    <div
+      id="collapseTwo"
+      class="accordion-collapse collapse"
+      data-bs-parent="#accordionExample"
+    >
+      <div class="accordion-body">
+        <strong>This is the second item's accordion body.</strong> It is hidden
+        by default, until the collapse plugin adds the appropriate classes that
+        we use to style each element. These classes control the overall
+        appearance, as well as the showing and hiding via CSS transitions. You
+        can modify any of this with custom CSS or overriding our default
+        variables. It's also worth noting that just about any HTML can go within
+        the <code>.accordion-body</code>, though the transition does limit
+        overflow.
+      </div>
+    </div>
+  </div>
+  <div class="accordion-item">
+    <h2 class="accordion-header">
+      <button
+        class="accordion-button collapsed"
+        type="button"
+        data-bs-toggle="collapse"
+        data-bs-target="#collapseThree"
+        aria-expanded="false"
+        aria-controls="collapseThree"
+      >
+        Accordion Item #3
+      </button>
+    </h2>
+    <div
+      id="collapseThree"
+      class="accordion-collapse collapse"
+      data-bs-parent="#accordionExample"
+    >
+      <div class="accordion-body">
+        <strong>This is the third item's accordion body.</strong> It is hidden
+        by default, until the collapse plugin adds the appropriate classes that
+        we use to style each element. These classes control the overall
+        appearance, as well as the showing and hiding via CSS transitions. You
+        can modify any of this with custom CSS or overriding our default
+        variables. It's also worth noting that just about any HTML can go within
+        the <code>.accordion-body</code>, though the transition does limit
+        overflow.
+      </div>
+    </div>
+  </div>
+</div>
